# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Deploy preview environments with ease",
        "title": "Buang API",
        "contact": {}
    },
    "basePath": "/api/v1",
    "paths": {
        "/diagnostics/info": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "api.v1",
                    "diagnostics"
                ],
                "summary": "Service info",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/diagnostics.InfoResult"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/diagnostics/stats": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Not the most accurate readings but it should tell which services need attention",
                "tags": [
                    "api.v1",
                    "diagnostics"
                ],
                "summary": "Docker stats",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/diagnostics.DockerStatsResult"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/project": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "api.v1",
                    "projects"
                ],
                "summary": "Find project by repository",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Repository",
                        "name": "repository",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/projects.FindProjectByRepositoryResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "api.v1",
                    "projects"
                ],
                "summary": "Create a project",
                "parameters": [
                    {
                        "description": "Project details",
                        "name": "projectDetails",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/projects.CreateProjectRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/project/{projectId}": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "api.v1",
                    "projects"
                ],
                "summary": "Update a project",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Project ID",
                        "name": "projectId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Project details",
                        "name": "projectDetails",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/projects.UpdateProjectRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "api.v1",
                    "projects"
                ],
                "summary": "Delete a project",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Project ID",
                        "name": "projectId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/project/{projectId}/branch": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "api.v1",
                    "projects"
                ],
                "summary": "Spin down a preview branch",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Project ID",
                        "name": "projectId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Branch details",
                        "name": "branchDetails",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/projects.BuangBranchRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/project/{projectId}/deployment": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "api.v1",
                    "deployments"
                ],
                "summary": "Spin up a preview deployment",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Project ID",
                        "name": "projectId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Deployment details",
                        "name": "deploymentDetails",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/deployments.CreateDeploymentRequest"
                        }
                    },
                    {
                        "type": "boolean",
                        "description": "Wait for deployment completion",
                        "name": "waitForDeployment",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/project/{projectId}/deployment/{deploymentId}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "api.v1",
                    "deployments"
                ],
                "summary": "Find deployment by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Project ID",
                        "name": "projectId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Deployment ID",
                        "name": "deploymentId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/deployments.FindDeploymentByIdResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "api.v1",
                    "deployments"
                ],
                "summary": "Spin down a preview deployment",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Project ID",
                        "name": "projectId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Deployment ID",
                        "name": "deploymentId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/project/{projectId}/deployment/{deploymentId}/logs": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "api.v1",
                    "deployment-logs"
                ],
                "summary": "Get deployment logs",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Project ID",
                        "name": "projectId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Deployment ID",
                        "name": "deploymentId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "Stream logs",
                        "name": "stream",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/project/{projectId}/deployments": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "api.v1",
                    "projects"
                ],
                "summary": "List all deployments",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Project ID",
                        "name": "projectId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/projects.ListAllDeploymentsItem"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/projects": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "api.v1",
                    "projects"
                ],
                "summary": "List all projects",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/projects.ListAllProjectsItem"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "deployments.CreateDeploymentRequest": {
            "type": "object",
            "required": [
                "branch",
                "serviceEntrypoint",
                "sha"
            ],
            "properties": {
                "branch": {
                    "type": "string"
                },
                "env": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "serviceEntrypoint": {
                    "type": "string"
                },
                "sha": {
                    "type": "string"
                }
            }
        },
        "deployments.FindDeploymentByIdResponse": {
            "type": "object",
            "properties": {
                "branch": {
                    "type": "string"
                },
                "deployedAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "projectId": {
                    "type": "integer"
                },
                "sha": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "diagnostics.ContainerStats": {
            "type": "object",
            "properties": {
                "cpuStats": {
                    "$ref": "#/definitions/diagnostics.CpuStats"
                },
                "image": {
                    "type": "string"
                },
                "imageId": {
                    "type": "string"
                },
                "memoryStats": {
                    "$ref": "#/definitions/diagnostics.MemoryStats"
                },
                "name": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "diagnostics.CpuStats": {
            "type": "object",
            "properties": {
                "usagePercent": {
                    "type": "number"
                }
            }
        },
        "diagnostics.DockerStatsResult": {
            "type": "object",
            "additionalProperties": {
                "$ref": "#/definitions/diagnostics.ContainerStats"
            }
        },
        "diagnostics.InfoResult": {
            "type": "object",
            "properties": {
                "timezone": {
                    "type": "string"
                },
                "uptime": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "diagnostics.MemoryStats": {
            "type": "object",
            "properties": {
                "limitMb": {
                    "type": "number"
                },
                "usageMb": {
                    "type": "number"
                },
                "usagePercent": {
                    "type": "number"
                }
            }
        },
        "projects.BuangBranchRequest": {
            "type": "object",
            "required": [
                "branch"
            ],
            "properties": {
                "branch": {
                    "type": "string"
                }
            }
        },
        "projects.CreateProjectRequest": {
            "type": "object",
            "required": [
                "composePath",
                "repository"
            ],
            "properties": {
                "composePath": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "repository": {
                    "type": "string"
                },
                "requiresAuthn": {
                    "type": "boolean"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "projects.FindProjectByRepositoryResponse": {
            "type": "object",
            "properties": {
                "composePath": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "repository": {
                    "type": "string"
                },
                "requiresAuthn": {
                    "type": "boolean"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "projects.ListAllDeploymentsItem": {
            "type": "object",
            "properties": {
                "deployedAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "projectId": {
                    "type": "integer"
                },
                "sha": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "projects.ListAllProjectsItem": {
            "type": "object",
            "properties": {
                "composePath": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "repository": {
                    "type": "string"
                },
                "requiresAuthn": {
                    "type": "boolean"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "projects.UpdateProjectRequest": {
            "type": "object",
            "required": [
                "composePath",
                "repository"
            ],
            "properties": {
                "composePath": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "repository": {
                    "type": "string"
                },
                "requiresAuthn": {
                    "type": "boolean"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        }
    }
}